added unit test for home paging source provider

+++ app/src/main/java/com/ba6ba/sadapaycasestudy/home/data/GitHubRepoResponse.kt
@@ -4,7 +4,7 @@ import com.google.gson.annotations.SerializedName
 
 data class GitHubRepoResponse(
     @SerializedName("total_count")
-    val totalCount: Long?,
+    val totalCount: Int?,
     @SerializedName("incomplete_results")
     val incompleteResults: Boolean?,
     val items: List<RepositoryItem>? = emptyList()
